首页接口对接，页面调整。

+++ app/src/main/java/com/shangyi/kt/fragment/home/model/HomeDataBean.kt
@@ -24,8 +24,8 @@ data class HomeDataBean(
 )
 
 data class Dealer(
-        val cash_back: Int,
-        val dealer: Int,
+        val cash_back: Float,
+        val dealer: Float,
         val gid: Int
 )
